add offline handling

+++ web-app/src/Redux/Actions/projects.js
@@ -6,19 +6,10 @@ import {
 } from "../types";
 
 import ProjectDataService from "../Services/ProjectService";
+import { toast } from 'react-toastify';
 
-export const createProject = (data) => async (dispatch) => {
-  try {
-    const res = await ProjectDataService.create(data);
-    dispatch({
-      type: CREATE_PROJECT,
-      payload: res.data.project,
-    });
-
-    return Promise.resolve(res.data);
-  } catch (err) {
-    return Promise.reject();
-  }
+const isOffline = () => {
+  return !navigator.onLine;
 };
 
 export const retrieveProjects = () => async (dispatch) => {
@@ -33,8 +24,30 @@ export const retrieveProjects = () => async (dispatch) => {
   }
 };
 
+export const createProject = (data) => async (dispatch) => {
+  try {
+    if(isOffline()) {
+      toast.error('You cannot perform this action while offline !');
+      return Promise.reject();
+    }
+    const res = await ProjectDataService.create(data);
+    dispatch({
+      type: CREATE_PROJECT,
+      payload: res.data.project,
+    });
+
+    return Promise.resolve(res.data);
+  } catch (err) {
+    return Promise.reject();
+  }
+};
+
 export const updateProject = (id, data) => async (dispatch) => {
   try {
+    if(isOffline()) {
+      toast.error('You cannot perform this action while offline !');
+      return Promise.reject();
+    }
     const res = await ProjectDataService.update(id, data);
     dispatch({
       type: UPDATE_PROJECT,
@@ -49,6 +62,10 @@ export const updateProject = (id, data) => async (dispatch) => {
 
 export const deleteProject = (id) => async (dispatch) => {
   try {
+    if(isOffline()) {
+      toast.error('You cannot perform this action while offline !');
+      return Promise.reject();
+    }
     const res = await ProjectDataService.remove(id);
     dispatch({
       type: DELETE_PROJECT,
